bdelmee/app/dataCollector/components/collector/requirements.txt
websocket-client
psycopg2==2.7.3.2
